Allow user to prefer static libraries in general
If the option -DUSE_STATIC=TRUE is set, then prefer static libraries
over shared ones, if they are present. However, if only a shared
library is present, then it will be used. I.e. the option only sets
the preference, it doesn't require anything since it is unlikely that
you have and want static linking of *everything*.