Require an OPM(-like) module to be linked statically
If -DOPM_CORE_USE_STATIC=TRUE is given in the configuration, then
only the static version of the opm-core library is searched for.
(See also the USE_STATIC option to indicate a general preference).
Even if the dynamic version is present, the search will fail if the
static libraries is not there.