EclipseWriter: implement writing the capillary pressure
I would not exactly call this code pretty, performant or memory
efficient, but given the current APIs it's probably as pretty,
performant and memory efficient as it can get. Note that this makes
the EclipseWriter more specific to the black-oil model as it requires
to pass and store a blackoil specific object to the constructor of the
EclipseWriter. As long as the state only features a single pressure
and the capillary pressure must be re-calculated by any piece of code
which needs it, I cannot see how to change this, though...